Split sensor.py into separate files for constants, data coordinator and sensor, add simple config flow
