+++ custom_components/frank_energie/strings.json
@@ -1,0 +1,9 @@
+{
+  "config": {
+    "step": {
+      "user": {
+          "description": "No authentication or configuration needed."
+      }
+    }
+  }
+}
