+++ custom_components/frank_energie/translations/nl.json
@@ -1,0 +1,9 @@
+{
+    "config": {
+        "step": {
+            "user": {
+                "description": "Geen authenticatie of configuratie nodig."
+            }
+        }
+    }
+}
